refactor: move dto into separate packages for passenger-service
