+++ passenger-service/.gitignore
@@ -31,3 +31,8 @@ build/
 
 ### VS Code ###
 .vscode/
+
+mvnw
+mvnw.cmd
+
+.mvn/
