+++ passenger-service/src/main/java/com/software/modsen/passengerservice/dto/request/PassengerRequest.java
@@ -1,4 +1,4 @@
-package com.software.modsen.passengerservice.dto;
+package com.software.modsen.passengerservice.dto.request;
 
 import lombok.Data;
 
@@ -7,5 +7,4 @@ public class PassengerRequest {
     private String name;
     private String email;
     private String phoneNumber;
-    private Double rating;
 }
