+++ passenger-service/src/main/java/com/software/modsen/passengerservice/dto/response/PassengerListDTO.java
@@ -1,4 +1,4 @@
-package com.software.modsen.passengerservice.dto;
+package com.software.modsen.passengerservice.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
