+++ passenger-service/src/main/java/com/software/modsen/passengerservice/mapper/PassengerMapper.java
@@ -1,7 +1,7 @@
 package com.software.modsen.passengerservice.mapper;
 
-import com.software.modsen.passengerservice.dto.PassengerRequest;
-import com.software.modsen.passengerservice.dto.PassengerResponse;
+import com.software.modsen.passengerservice.dto.request.PassengerRequest;
+import com.software.modsen.passengerservice.dto.response.PassengerResponse;
 import com.software.modsen.passengerservice.model.Passenger;
 import org.mapstruct.Mapper;
 import org.mapstruct.MappingTarget;
