+++ passenger-service/src/main/java/com/software/modsen/passengerservice/service/PassengerService.java
@@ -1,8 +1,8 @@
 package com.software.modsen.passengerservice.service;
 
-import com.software.modsen.passengerservice.dto.PassengerListDTO;
-import com.software.modsen.passengerservice.dto.PassengerRequest;
-import com.software.modsen.passengerservice.dto.PassengerResponse;
+import com.software.modsen.passengerservice.dto.response.PassengerListDTO;
+import com.software.modsen.passengerservice.dto.request.PassengerRequest;
+import com.software.modsen.passengerservice.dto.response.PassengerResponse;
 
 public interface PassengerService {
     PassengerListDTO getPassengers();
